插件修改

See merge request rapserver/smarthelper!1019

+++ common-lib/src/main/java/com/zfs/common/pojo/PluginPO.java
@@ -38,17 +38,17 @@ public class PluginPO implements Serializable {
 
     private String url;
     //插件版本，zfs新增
-    private Byte pluginv;
+    private Long pluginv;
     //插件包名
     private String pluginpkg;
     //插件类型
     private Byte type;
 
-    public Byte getPluginv() {
+    public Long getPluginv() {
         return pluginv;
     }
 
-    public void setPluginv(Byte pluginv) {
+    public void setPluginv(Long pluginv) {
         this.pluginv = pluginv;
     }
